+++ core/src/main/scala/me/shadaj/slinky/core/facade/Fragment.scala
@@ -1,0 +1,7 @@
+package me.shadaj.slinky.core.facade
+
+import me.shadaj.slinky.core.ExternalComponentNoProps
+
+object Fragment extends ExternalComponentNoProps {
+  override val component = React.Fragment
+}