+++ core/src/main/scala/me/shadaj/slinky/core/facade/React.scala
@@ -24,6 +24,8 @@ object React extends js.Object {
   class Component(jsProps: js.Object) extends js.Object {
     def forceUpdate(callback: js.Function0[Unit]): Unit = js.native
   }
+
+  val Fragment: js.Object = js.native
 }
 
 @js.native