+++ web/src/main/scala/me/shadaj/slinky/web/ReactDOM.scala
@@ -10,5 +10,12 @@ import scala.scalajs.js.annotation.JSImport
 @JSImport("react-dom", JSImport.Namespace, "ReactDOM")
 object ReactDOM extends js.Object {
   def render(component: ReactElement, target: Element): ReactInstance = js.native
+  def hydrate(component: ReactElement, target: Element): ReactInstance = js.native
   def findDOMNode(instance: React.Component): Element = js.native
+}
+
+@js.native
+@JSImport("react-dom/server", JSImport.Namespace, "ReactDOMServer")
+object ReactDOMServer extends js.Object {
+  def renderToString(element: ReactElement): String = js.native
 }